cloud-function-audio/requirements.txt
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1